feat: Create "config.py" moude for "models" package.
Add "__repr__()" for "Base" class.

Create "pk_tp", "time_update_tp", "time_create_tp" often-used types in Modles.
To reduce code duplication.

refactor: Change project structure.
