+++ backend/app/db/__init__.py
@@ -1,0 +1,4 @@
+from . import models
+from .config import Session, create_db_tables, engine
+
+__all__ = ["Session", "create_db_tables", "engine", "models"]
