+++ backend/app/db/models/__init__.py
@@ -1,0 +1,16 @@
+
+from .application import Application, AppStatus, Company, WorkLocation, WorkType
+from .config import Base, pk_tp, time_create_tp, time_update_tp
+
+__all__ = [
+    "AppStatus",
+    "Application",
+    "Base",
+    "Company",
+    "WorkLocation",
+    "WorkType",
+    "pk_tp",
+    "time_create_tp",
+    "time_update_tp",
+]
+
